+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.9.2 (unreleased)
+## 0.10.0 (2015-02-28)
 
   * renamed the `localizerDestDir` option to `localizerOutputDir`, changed its type to `Object` and fixed the
     configuration to recognize a non-default value